Renamed Context of the Golden/Movements module and fixed the namespaces of it
