+++ Assets/Modules/NPCs/Golden/Movements/MovementContext.cs.meta
